requirements.txt
# Trading Strategy Dependencies - Modern Architecture
# Install with: pip install -r requirements.txt

# Core data and analysis
pandas>=2.3.0
numpy>=2.2.0
python-binance>=1.0.29
pytz>=2025.2

# API and real-time server
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Optional: For advanced plotting (legacy support)
plotly>=6.2.0

# Additional dependencies (automatically installed)
requests>=2.32.0
aiohttp>=3.12.0
websockets>=15.0.0
pycryptodome>=3.23.0
six>=1.17.0
python-dateutil>=2.9.0
